# [Enhancement]:Optimize bitmap index writer