Signed-off-by: xyz <a997647204@gmail.com>

## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #5328 

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

In the current BitmapIndexWriterImpl code, add_value needs to call twice getSizeInBytes the monitor the change of the roaring map. The add_value function is called very frequently. Which has a great impact on the performance of compaction in the test.
This pr optimizes the calculation of roaring size in the bitmap Index. 

The solution to this problem is to move the calculation of the roaring bitmap size from the `add_value` function` to the `size` function. Delay calculation.