@xiaoyong-z 2022-04-21
Base compaction flame graph before optimization:
![image](https://user-images.githubusercontent.com/31178013/164356218-5d4e3df5-d406-4295-b7bc-646d760001bc.png)
Base compaction flame graph after optimization:
![image](https://user-images.githubusercontent.com/31178013/164356276-201d841a-8b98-4671-ba9e-eda3253ca9e9.png)