@xiaoyong-z 2022-04-21
> I have some questions
> 
> 1. is the getSizeInBytes cost dominate? or the _mem_index.find(). If all the values are different, each item will call ** getSizeInBytes** only once but the _mem_index.find() twice. if all the values are the same it will call getSizeInBytes twice but call _mem_index.find() once.
> 2. if the size() is called many times, getSizeInBytes will be called many times. and it won't have a better performance.
> 
> If the _mem_index.find() is the dominate, I think it is easy to eliminate it.

1. In the flame graph, it shows that `getSizeInBytes` dominates.
2. i will sumbit a optimized pr later to reduce the call to minimize calls to getSizeInBytes more.