@imay 2022-04-21
> 1. In the flame graph, it shows that `getSizeInBytes` dominates.

It makes sense. I recheck your test case again and found that you use tiny int, which will have a lot of duplicates values. Then there will be too many getSizeInBytes calls.
So I think you can test the worst cases, which there are no same value at all. then your change may introduce more getSizeInBytes calls if size() is called many times. For example you create a bitmap index on the int64_t column, whose values are different with each other.