@xiaoyong-z 2022-04-24
# New experimental results
## Table
CREATE TABLE `table1` (
  `key1` int NOT NULL COMMENT "",
  `key2` datetime NOT NULL COMMENT "",
  `key3` tinyint NOT NULL COMMENT "",
  `key4` smallint NOT NULL COMMENT "",
  `key5` tinyint NOT NULL COMMENT "",
  `key6` tinyint NOT NULL COMMENT "",
  `key7` tinyint NOT NULL COMMENT "",
  `key8` tinyint NOT NULL COMMENT "",
  `key9` smallint NULL COMMENT "",
  `key10` varchar(20) NOT NULL DEFAULT "baishan" COMMENT "",
  `key11` datetime NOT NULL COMMENT "",
  `key12` datetime NOT NULL COMMENT "",
  `key13` datetime NOT NULL COMMENT "",
  `value1` bigint SUM NULL COMMENT "",
  `value2` bigint SUM NULL COMMENT "",
  `value3` bigint SUM NULL COMMENT "",
  `value4` bigint SUM NULL DEFAULT "0" COMMENT "",
  `update_time` datetime REPLACE NULL COMMENT "",
) ENGINE=OLAP
AGGREGATE KEY(`key1`, `key2`, `key3`, `key4`, `key5`, `key6`, `key7`, `key8`, `key9`, `key10`, `key11`, `key12`, `key13`)
COMMENT "OLAP"
DISTRIBUTED BY HASH(`key1`) BUCKETS 1
PROPERTIES (
"replication_num" = "1",
"in_memory" = "false",
"storage_format" = "V2"
);
## Design four sets of comparison experiments
1. No bitmap index
2. bitmap index on tinyint(-128,127)
```sql
INDEX idx_bitmap_key3 (`key3`) USING BITMAP COMMENT 'key3 bitmap idx' 
```
3. bitmap index on Smallint(-32768, 32767)
```sql
INDEX idx_bitmap_key4 (`key4`) USING BITMAP COMMENT 'key4 bitmap idx'
```
4. bitmap index on Int(-2147483648, 2147483647)
```sql
INDEX idx_bitmap_key1 (`key1`) USING BITMAP COMMENT 'key1 bitmap idx'
```
## Experiment 1
base compaction time 150s
## Experiment 2
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 259.837s | 1.2G | 109.837 | 1x
After Optimization | 168.927s | 1.2G | 18.027 | 6.09x

</div>
## Experiment 3
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 611.759s | 1.33G | 461.759s | 1x
After Optimization | 204s | 0.31G | 54s | 8.55s

</div>
## Experiment 4
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 267s | 2.53G | 117s | 1x
After Optimization | 226.449s | 0.534G | 76.449s | 1.53x

</div>