@xiaoyong-z 2022-04-24
# Principle of Roaring map
Following is the struct of roaring map
![image](https://user-images.githubusercontent.com/31178013/164977448-ae4c297a-ed61-494f-9ce7-6422e657b62f.png)
left is the key, it is the higher 16 bits of the u32_t
right is the container, there are three types of containers:
ArrayContainer, Bitmap Container, RunContainer.
![image](https://user-images.githubusercontent.com/31178013/164977508-7f023c6e-7aa2-4ac0-81f6-930debef4d72.png)
Their relationship:
![image](https://user-images.githubusercontent.com/31178013/164977532-be0282a4-7cfa-46a5-ad4e-ac63d3488cc1.png).
Reference https://zhuanlan.zhihu.com/p/445396980
# Design idea