@xiaoyong-z 2022-04-24
# Design idea
When the number of keys in the roaring bitmap is relatively moderate (less than 4096, more than 512), there will be a lot of array containers with only 1 or 2 elements (while in fact each of these array containers can hold 4096 elements). Each of these containers can hold 4096 elements), creating these array containers is both very time consuming and a huge memory waste, as we saw in Experiment 3.
So my idea is to manage the elements ourselves when roaring bitmap with 1024 elements and flush them as CROARING_SERIALIZATION_ARRAY_UINT32 at flush time