@xiaoyong-z 2022-04-25
# Experimental results After removing  hardcode roaring array
## Experiment 1
base compaction time 150s
## Experiment 2
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 259.837s | 1.2G | 109.837 | 1x
After Optimization | 163.927s | 1.29G | 13.027 | 8.43x

</div>
## Experiment 3
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 611.759s | 1.33G | 461.759s | 1x
After Optimization | 295.5s | 1.39G | 145.5s | 3.17x

</div>
## Experiment 4
<div class="okr-block-clipboard" data-okr="%7B%22okrDelta%22%3A%5B%7B%22lineType%22%3A%22unsupport%22%2C%22lineOptions%22%3A%7B%7D%2C%22lineContent%22%3A%5B%5D%7D%5D%2C%22businessKey%22%3A%22lark-doc%22%7D"></div><div data-zone-id="0" data-line-index="0" style="white-space: pre;">

  | Base compaction Time | Maximum memory | Index builder time | Speed Up
-- | -- | -- | -- | --
Before Optimization | 267s | 2.53G | 117s | 1x
After Optimization | 254s | 2.71G | 104s | 1.125x

</div>